==> src/4) data analysis/1) Cascading.ipynb <==
import pandas as pd
import swifter
import numpy as np

df = pd.read_csv('../../data/all_cleaned.csv')

df = df[df['Time'] > 1540947600]

df['cascade'] = df['in_response_to']

df.loc[df[df['cascade'] == 0].index, 'cascade'] = df[df['cascade'] == 0].ID #set cascade as ID if cascade = 0

len(df)

vals = df['cascade'].value_counts().reset_index()

df = df[df['cascade'].isin(vals[vals['cascade'] > 1]['index'])]

len(df)

df.to_csv('data/retweetcascade.csv', index=None)

len(df.groudfpby('cascade'))

vals[vals['cascade'] > 2]
