==> src/4) data analysis/2) Testing Cascades.ipynb <==
import pandas as pd
import swifter

cascaded = pd.read_csv('data/curr_influenceretweet_cascade.csv', header=None)

filename = 'retweetcascade'
df = pd.read_csv('data/{}.csv'.format(filename))

df = df[~df['cascade'].isin(list(set(cascaded[2])))]

len(df)

df = pd.read_csv('data/retweetcascade.csv')

original = pd.read_csv('../../data/all_cleaned.csv')

parents = original[original['in_response_to'] == 0]

parents = parents[parents['ID'].isin(df['in_response_to'])]

parents[parents['ID'].isin(df['ID'])] #they should be here. Mistake in cascading

analyzing = pd.merge(original[['ID', 'in_response_to']], df[['ID', 'cascade', 'in_response_to']], on='ID', suffixes=('_original', '_cascaded'))


len(analyzing)

not_same = analyzing[analyzing['in_response_to_original'] != analyzing['cascade']]

not_same[not_same['in_response_to_original'] != not_same['in_response_to_cascaded']] #okay

not_same = not_same.drop('in_response_to_cascaded', axis=1)

not_same = not_same[not_same['in_response_to_original'] != 0] #zero are okay from visual inspection
