==> src/4) data analysis/3) Retweet Influence Calculation.ipynb <==
import pandas as pd
from casIn.user_influence import P,influence
import swifter
from glob import glob
import numpy as np
import time

import signal

class timeout:
    def __init__(self, seconds=1, error_message='Timeout'):
        self.seconds = seconds
        self.error_message = error_message
    def handle_timeout(self, signum, frame):
        raise TimeoutError(self.error_message)
    def __enter__(self):
        signal.signal(signal.SIGALRM, self.handle_timeout)
        signal.alarm(self.seconds)
    def __exit__(self, type, value, traceback):
        signal.alarm(0)

filename = 'retweetcascade'
colname = 'cascade'
count=0

def get_influence(df):
    global count
    global colname
    count = count + 1

    df = df.reset_index(drop=True)

    p_ij = P(df,r = -0.000068)
    inf, m_ij = influence(p_ij)
    df['inf'] = inf
    df = df[['ID', 'inf', colname]]
    df.to_csv('data/curr_influenceretweet_cascade.csv', mode='a', header=None, index=None)
    print(count)

df = pd.read_csv('data/{}.csv'.format(filename))

len(df)

# server_cascaded = pd.read_csv('data/server_curr_influenceretweet_cascade.csv', header=None)
# cascaded = pd.read_csv('data/curr_influenceretweet_cascade.csv', header=None)
# cascaded = pd.concat([cascaded, server_cascaded]).drop_duplicates().reset_index(drop=True)
# cascaded.to_csv('data/curr_influenceretweet_cascade.csv', index=None)

cascaded = pd.read_csv('data/curr_influenceretweet_cascade.csv', header=None)
df = df[~df['cascade'].isin(list(set(cascaded[2])))]

len(df)

# df.to_csv('remaining.csv')

req_cols = ['ID','Time', 'followers_count', 'User', 'predicted', 'vader_emotion', 'polarity', 'subjectivity', 'Positive', 'Negative']
req_cols.append(colname)
df = df[req_cols]
df = df.rename(columns={'Time': 'time', 'followers_count': 'magnitude', 'User': 'user_id'})

df.groupby(colname).apply(get_influence)

df = pd.read_csv('data/{}.csv'.format(filename))
d = pd.read_csv('data/curr_influenceretweet_cascade.csv').rename(columns={'0': 'ID', '1': 'influence', '2': 'cascade'})
d = d.drop_duplicates()

df=df.merge(d, on='ID')
df = df.drop('cascade_y', axis=1).rename(columns={'cascade_x':'cascade'})
df = df[df['Time'] > 1540947600]
df.to_csv('cascade_with_influence.csv', index=None)

df
